+++ src/assets/gulp/tasks/build.js
@@ -3,7 +3,6 @@
  */
 const display = require('./util/display');
 const template = require('./util/template');
-const modules = require('./util/modules');
 const uc = require('./util/unite-config');
 const buildTranspile = require('./build-transpile');
 const buildBundle = require('./build-bundle');
@@ -15,7 +14,7 @@ const runSequence = require('run-sequence');
 
 gulp.task('build-clean', (callback) => {
     const uniteConfig = uc.getUniteConfig();
-    const toClean = path.resolve(uniteConfig.directories.dist);
+    const toClean = path.resolve(uniteConfig.directories.dist) + "**/!(app-module-config).js";
     display.info('Cleaning', toClean);
     return del(toClean, callback);
 });
@@ -24,19 +23,12 @@ gulp.task('build-generate-index', () => {
     display.info('Generating', "index.html from index.html.template");
 
     const uniteConfig = uc.getUniteConfig();
-    const moduleConfig = path.join(uniteConfig.directories.dist, "module-config.js")
+    const moduleConfig = path.join(uniteConfig.directories.dist, "app-module-config.js")
 
     return template.copyTemplate("index.html.template", "index.html", moduleConfig.replace(/\\/g, '/'), uniteConfig, "node_modules");
 });
 
-gulp.task('build-generate-module-config', (cb) => {
-    const uniteConfig = uc.getUniteConfig();
-    
-    display.info('Generating Module Configuration', "module-config.js");
-    modules.createModuleConfig(path.join(uniteConfig.directories.dist, "module-config.js"), "index.html", uniteConfig, "node_modules", cb);
-});
-
 gulp.task('build', (cb) => {
-    runSequence('build-clean', 'build-transpile', 'build-bundle', 'build-generate-index', 'build-generate-module-config', cb);
+    runSequence('build-clean', 'build-transpile', 'build-bundle', 'build-generate-index', cb);
 });