+++ src/configuration/models/unite/includeMode.ts
@@ -1,0 +1,4 @@
+/**
+ * Include modes.
+ */
+export type IncludeMode = "app" | "test" | "both";