+++ src/interfaces/IPackageManager.ts
@@ -1,8 +1,10 @@
 /**
  * Interface for package manager.
  */
+import { PackageConfiguration } from "../configuration/models/packages/packageConfiguration";
+
 export interface IPackageManager {
-    latestVersion(packageName: string): Promise<string>;
-    add(packageName: string, version: string, isDev: boolean): Promise<void>;
-    remove(packageName: string, isDev: boolean): Promise<void>;
+    info(packageName: string): Promise<PackageConfiguration>;
+    add(workingDirectory: string, packageName: string, version: string, isDev: boolean): Promise<void>;
+    remove(workingDirectory: string, packageName: string, isDev: boolean): Promise<void>;
 }