+++ src/pipelineSteps/moduleLoader.ts
@@ -16,9 +16,18 @@ export class ModuleLoader extends EnginePipelineStepBase {
             if (uniteConfiguration.moduleLoader === "RequireJS") {
                 engineVariables.requiredDependencies.push("requirejs");
                 uniteConfiguration.staticClientModules.push("requirejs/require.js");
+
+                uniteConfiguration.srcDistReplace = "(define)*?(..\/src\/)";
+                uniteConfiguration.srcDistReplaceWith = "../dist/";
             } else if (uniteConfiguration.moduleLoader === "SystemJS") {
                 engineVariables.requiredDependencies.push("systemjs");
                 uniteConfiguration.staticClientModules.push("systemjs/dist/system.js");
+
+                uniteConfiguration.srcDistReplace = "(System.register)*?(..\/src\/)";
+                uniteConfiguration.srcDistReplaceWith = "../dist/";
+            } else if (uniteConfiguration.moduleLoader === "Webpack" || uniteConfiguration.moduleLoader === "Browserify") {
+                uniteConfiguration.srcDistReplace = "(require)*?(..\/src\/)";
+                uniteConfiguration.srcDistReplaceWith = "../dist/";
             }
 
             return 0;